fix: refactor auth modal
