+++ apps/mochi-web/components/AuthWidget/LoginRequireModal.tsx
@@ -6,8 +6,7 @@ import {
 import { useAuthStore } from "~store"
 import clsx from "clsx"
 import { useEffect } from "react"
-import { AuthPanel } from "./auth-panel"
-import { useRouter } from "next/router"
+import { AuthPanel } from "./AuthPanel"
 
 export const LoginRequiredModal = () => {
